# backend/api/openapi.yaml
...
openapi: 3.1.0
...
info:
  title: Foraged goods marketplace API
  version: 0.1.0
  # "|" keeps the newlines - the description renders as Markdown in the UI.
  description: |
    A marketplace for foraged goods — berries, mushrooms, herbs. Sellers post
    listings, buyers browse, message them and place orders. Payments are
    simulated; no money changes hands.

    ## Authentication

    Two tokens, with opposite properties.

    The **access token** is a signed JWT sent on every request:

        Authorization: Bearer <access_token>

    It is verified without touching the database, which also means it cannot be
    revoked - so it is short-lived (15 minutes by default, `ACCESS_TOKEN_TTL`).

    The **refresh token** is an opaque value in an HttpOnly `refresh_token`
    cookie, set by signup and login and valid for 7 days. `POST /auth/refresh`
    exchanges it for a new access token and rotates it. Only its hash is
    stored, so a database leak cannot resume a session.

    When an access token expires, the 401 carries
    `WWW-Authenticate: Bearer error="invalid_token"` - the signal to refresh
    rather than to show a login screen. `GET /auth/me` returns the identity
    behind a token that is still valid, for restoring a session on page load.

    Public endpoints (browsing listings, search, images) work without one.

    ## Ids

    Every id is a **uuid v7**: the leading 48 bits are a millisecond timestamp,
...
    something to parse.

    ## Errors

...

...

...
servers:
...

...
security:
  - bearerAuth: []
  - apiKeyAuth: []

# "components" is the reusable-definitions section. Nothing in here appears in
# the docs on its own - it only shows up where something $refs it.
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        A key from `POST /me/api-keys`, for machine clients. It acts as its
        owner - same permissions, same ownership checks - and is rate limited
        per key rather than per IP.

        Keys cannot manage keys: creating, listing and revoking need a session,
        so a leaked key cannot mint a replacement and survive its own
        revocation.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # documentation only; no tool verifies the format
      description: |
        Access token issued by signup or login. Expired and missing tokens are
        both answered with 401.
  schemas:
...
    Error:
      type: object
      required: [error] # "details" is omitempty in Go, so it is not required
      properties:
        error:
          type: string
          description: |
            Safe to show a user on 4xx, and written for display: sentence case,
            no trailing full stop. Render it as-is rather than transforming it.
            On 5xx this is always "Something went wrong" — the cause is in the
            server log, tagged with the `X-Request-ID` on the response.
          examples: ["Listing not found"]
        details:
          type: object
          # "any key, string value" - the YAML spelling of map[string]string.
          additionalProperties:
            type: string
          description: Reserved for field-level validation errors; absent today.
    ListingImage:
      type: object
      required: [id, url, position]
      properties:
        id:
          type: string
          format: uuid
          examples: ["01a02305-b81c-7dcb-86a0-7f75e33e0af3"]
        url:
          type: string
          description: Path served by the API, e.g. for use in an `img` tag.
          examples: ["/uploads/0a7a0bc6-99f9-42f8-9ba0-c908a9cb3a93.png"]
        position:
          type: integer
          format: int32
          description: Display order, ascending. Unique per listing.
          examples: [0]
    Listing:
      type: object
      # Every field without omitempty in ListingResponse is always present.
      required:
        [
          id,
          seller_id,
          title,
          description,
          category,
          price,
          quantity,
          unit,
          created_at,
          updated_at,
          images,
        ]
      properties:
        id:
          type: string
          format: uuid
          examples: ["01a02305-b81c-7dcb-86a0-7f75e33e0af3"]
        seller_id:
          type: string
          format: uuid # Go uuid.UUID marshals to a string
        title:
          type: string
          maxLength: 100
          examples: ["Golden Chanterelles"]
        description:
          type: string
          description: Empty string when the seller left it blank.
        category:
          type: string
          examples: ["mushrooms"]
        price:
          type: number
          format: double
          description: Price per unit. Stored as NUMERIC(10,2).
          examples: [18.5]
        quantity:
          type: integer
          format: int32
          description: Remaining quantity; 0 means sold out.
          examples: [4]
        unit:
          type: string
          examples: ["kg"]
        created_at:
          type: string
          format: date-time # Go time.Time marshals to RFC 3339
        updated_at:
          type: string
          format: date-time
        images:
          type: array
          description: Always present; empty when the listing has no photos.
          items:
            # $ref = "reuse that definition here". This is why schemas exist.
            $ref: "#/components/schemas/ListingImage"
    PaginatedListings:
      type: object
      required: [items, total, page, limit, total_pages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
          format: int64
          description: Matching listings across all pages.
        page:
          type: integer
          description: Echoes the requested page, even past the last one.
        limit:
          type: integer
          description: Clamped to 50.
        total_pages:
          type: integer
...
    ListingInput:
      type: object
      # Create and update take the identical body, so they share this schema.
      # PUT is a full replace: every field is required on both.
      required: [title, category, price, quantity, unit]
      properties:
        title:
          type: string
          maxLength: 100
          description: |
            The server counts BYTES, so accented characters use up more than
            one of the 100.
          examples: ["Golden Chanterelles"]
        description:
          type: string
          description: The only optional field.
        category:
          type: string
          maxLength: 50
          examples: ["mushrooms"]
        price:
          type: number
          format: double
          exclusiveMinimum: 0 # "price must be greater than 0"
          examples: [18.5]
        quantity:
          type: integer
          format: int32
          exclusiveMinimum: 0
          examples: [4]
        unit:
          type: string
          maxLength: 20
          examples: ["kg"]
...
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json: # the media type the body is sent as
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: |
        No token, or one that is expired or invalid. The body is the same in
        every case; `WWW-Authenticate` says which, because a client's response
        differs: refresh, or log in.
      headers:
        WWW-Authenticate:
          description: |
            `Bearer error="invalid_token", error_description="The access token expired"`
            when the token was valid but has aged out - retry through
            `POST /auth/refresh`. A bare `Bearer` for a missing, malformed or
            forged token - the client should log in. A forged token reads as
            absent on purpose.
          schema:
            type: string
          examples:
            expired:
              value: Bearer error="invalid_token", error_description="The access token expired"
            absent:
              value: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: |
        Authenticated, but not permitted — the resource belongs to someone
        else, or the account does not hold the role the route requires.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: |
        No such resource — also returned when a resource exists but the caller
        is not a participant, so its existence is not revealed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: |
        Too many requests on this API key. Only key-authenticated requests are
        limited; browser sessions are not.
      headers:
        Retry-After:
          description: Seconds until a request will be accepted again.
          schema:
            type: integer
            examples: [2]
        X-RateLimit-Limit:
          description: |
            The bucket's capacity. Sent on every key-authenticated response,
            not only on a refusal.
          schema:
            type: integer
            examples: [60]
        X-RateLimit-Remaining:
          description: Requests left in the current bucket.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: Something failed on our side; the cause is in the log.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

# Each key is a URL relative to the server above, so "/listings" is really
# http://localhost:8080/api/v1/listings.
paths:
...
  /listings:
    get:
      tags: [listings]
      summary: List the available listings
      operationId: getListings
      security: []
      description: |
        Newest first, and **sold-out listings are excluded** (`quantity > 0`).
        A sold-out listing is still reachable through `/listings/{id}` and
        still appears in `/me/saved`.

        Unpaginated — use `/listings/search` for filtering, sorting and pages.
      responses:
        "200":
          description: The available listings, each with its photos attached.
          content:
            application/json:
              schema:
                # An array response has no named schema; describe it inline.
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"

    post:
      tags: [listings]
      summary: Create a listing
      operationId: createListing
      # No "security" override here: the document default (bearer token)
      # applies, which is exactly what this endpoint needs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: |
            Created. `images` is always `[]` here — upload photos afterwards
            with `POST /listings/{id}/images`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /listings/search:
    get:
      tags: [listings]
      summary: Search, filter, sort and paginate listings
      operationId: searchListings
      security: []
      description: |
        Every parameter is optional; omitting one means "no filter". Sold-out
        listings are excluded, as in `GET /listings`.

        `keyword` and `location` match case-insensitively on a substring, with
        `%` or `_` in the input treated literally rather than as SQL
        wildcards. `category` is an exact, case-SENSITIVE match.
      parameters:
        # in: query = ?keyword=... . Optional unless "required: true".
        - name: keyword
          in: query
          description: Matches the title or description.
          schema:
            type: string
            maxLength: 200
          examples:
            chanterelle:
              value: chanterelle
        - name: category
          in: query
          schema:
            type: string
            maxLength: 200
          examples:
            mushrooms:
              value: mushrooms
        - name: location
          in: query
          description: |
            Matches the seller's address, not the listing — a listing whose
            seller has no address is never returned by this filter.
          schema:
            type: string
            maxLength: 200
        - name: min_price
          in: query
          description: Non-negative. Must not exceed `max_price`.
          schema:
            type: number
            minimum: 0
        - name: max_price
          in: query
          schema:
            type: number
            minimum: 0
        - name: sort
          in: query
          description: Anything outside this list is a 400.
          schema:
            type: string
            # Mirrors the sortOptions map in database/listings_search.go.
            enum: [newest, oldest, price_asc, price_desc]
            default: newest
        - name: page
          in: query
          description: 1-based. A page past the end returns an empty `items`.
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Values above 50 are clamped to 50 rather than rejected.
          schema:
            type: integer
            minimum: 1
            default: 20
      responses:
        "200":
          description: One page of matches, plus the totals for a pager.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedListings"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /listings/{id}:
    # Declared at path level, so every method on this path inherits it.
    # {id} in the path and "name: id" here have to match exactly.
    parameters:
      - name: id
        in: path
        required: true # path parameters are ALWAYS required
        schema:
          type: string
          format: uuid
    get:
      tags: [listings]
      summary: Get one listing
      operationId: getListing
      security: []
      responses:
        "200":
          description: The listing, with its photos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"

    put:
      tags: [listings]
      summary: Replace a listing
      operationId: updateListing
      description: |
        A full replace, not a patch: every required field must be sent, and an
        omitted `description` is stored as empty. Sellers only — and a listing
        whose quantity has reached 0 can no longer be edited.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "200":
          description: The updated listing, with its photos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The listing is sold out; create a new one instead.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [listings]
      summary: Delete a listing
      operationId: deleteListing
      description: |
        Refused once the listing has orders — the order history has to keep
        pointing at something, so the foreign key is ON DELETE RESTRICT.
      responses:
        "204":
          description: Deleted. No body.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The listing has orders and cannot be deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /listings/{id}/images:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [images]
      summary: List a listing's photos
      operationId: getListingImages
      security: []
      description: |
        Ordered by `position`. A listing id that does not exist returns an
        empty array, not a 404.
      responses:
        "200":
          description: The photos, ascending by position.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListingImage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"

    post:
      tags: [images]
      summary: Upload a photo
      operationId: uploadListingImage
      description: |
        `multipart/form-data` with one file field named `image`. The type is
        decided by sniffing the file's first 512 bytes, not by the filename or
        the client's Content-Type — so a `.png` that is really a PDF is a 415.

        Limits come from the environment: `MAX_UPLOAD_BYTES` (default 5 MiB)
        and `MAX_IMAGES_PER_LISTING` (default 5).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary # "this field is a file", not text
      responses:
        "201":
          description: |
            Stored. `position` is one past the highest already in use, so a
            gap left by a deleted photo is never reused.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingImage"
        "400":
          description: Not a multipart form, or no `image` field in it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: This listing already has the maximum number of images.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: The upload exceeded `MAX_UPLOAD_BYTES`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: Not a JPEG, PNG or WebP.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"

  # ---------------------------------------------------------------------------
  # served outside /api/v1 - hence the per-path "servers" override below
  # ---------------------------------------------------------------------------
...
  /listings/{id}/images/{imageID}:
    # Two path parameters. Note the casing: chi's route says {imageID}, so the
    # spec has to say {imageID} too.
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: imageID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [images]
      summary: Delete a photo
      operationId: deleteListingImage
      description: |
        Removes the row and the file. Remaining photos keep their positions;
        gaps are allowed.
      responses:
        "204":
          description: Deleted. No body.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"

  # ---------------------------------------------------------------------------
  # wishlist
  # ---------------------------------------------------------------------------
...